refactored types so that loader indicator extends types
