+++ packages/components/src/Loader/Loader.tsx
@@ -1,47 +1,9 @@
 import React, { PropsWithChildren, RefObject } from "react";
 import classnames from "classnames";
-import { XOR } from "ts-xor";
 import { Breakpoints, useResizeObserver } from "@jobber/hooks/";
+import { LoaderProps } from "./LoaderProps";
 import styles from "./Loader.css";
 import { LoaderIndicator } from "./LoaderIndicator";
-
-interface BaseLoaderProps {
-  /**
-   * Removes the overlay and has the component render in the flow of the layout
-   * @default false
-   */
-  readonly inline?: boolean;
-
-  /**
-   * Shows a spinner when false, progress bar when true.
-   * @default false
-   */
-  determinate?: boolean;
-
-  /**
-   * Show loading wrapper if underlying content is loading.
-   * @default false
-   */
-  loading: boolean;
-}
-
-interface DeterminateLoaderProps extends BaseLoaderProps {
-  determinate: true;
-
-  /**
-   * The determinate loading progress
-   * @default 0
-   */
-  currentValue: number;
-
-  /**
-   * The determinate loading value when loading is completed
-   * @default 100
-   */
-  maxValue?: number;
-}
-
-type LoaderProps = XOR<BaseLoaderProps, DeterminateLoaderProps>;
 
 export function Loader(props: PropsWithChildren<LoaderProps>) {
   const { loading = false, children } = props;
