+++ packages/components/src/Loader/LoaderIndicator.tsx
@@ -1,13 +1,12 @@
 import React from "react";
 import styles from "./LoaderIndicator.css";
+import { LoaderProps } from "./LoaderProps";
 import { Spinner } from "../Spinner";
 import { ProgressBar } from "../ProgressBar";
 
-interface LoaderIndicatorProps {
-  determinate: boolean;
+interface LoaderIndicatorProps
+  extends Pick<LoaderProps, "determinate" | "currentValue" | "maxValue"> {
   size: "base" | "small";
-  currentValue?: number;
-  maxValue?: number;
 }
 
 export function LoaderIndicator({
